Fix rpc communication (#40)
* Adjust rpc calls on the elixir side

* Actually poll periodically

* 0x before serializing

* Better params

* 0x before serializing

* revert changes

* rpc fix

* Added block querying by hash

* Extract stripping code (#28)

* Extract stripping code

* fix comment

* feature flag

* undo rpc change

* Fix api

* rpc fix

* fix merge conflicts + fmt clippy

* 0x (de)serializatoin fix

* fmt clippy

* small fix

* Fix poller to start from the beginning and go upwards, add proofs directory to .gitignore

* latest block tag

* transaction receipt

* BlockId::Tag(Latest) is now just BlockId::Latest

* pass reference

* timestamp as milliseconds in blocks

* timestamp as u128

* u128 fix

* remove lossy string deserialization

* revert type tag

* block deserialization

* Fix compilation

---------

+++ sequencer/rpc_endpoint/src/rpc/types/codegen.rs
@@ -33,11 +33,12 @@ use serializable_types::NumAsHex;
 /// Block status.
 ///
 /// The status of the block.
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize, Deserialize, Hash)]
+#[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize, Deserialize, Hash, Default)]
 pub enum BlockStatus {
     #[serde(rename = "PENDING")]
     Pending,
     #[serde(rename = "ACCEPTED_ON_L2")]
+    #[default]
     AcceptedOnL2,
     #[serde(rename = "ACCEPTED_ON_L1")]
     AcceptedOnL1,
@@ -90,7 +91,7 @@ pub struct BlockWithTxHashes {
 ///
 /// The block object.
 #[serde_as]
-#[derive(Debug, Clone, Serialize, Deserialize)]
+#[derive(Debug, Clone, Serialize, Default, Deserialize)]
 #[cfg_attr(feature = "no_unknown_fields", serde(deny_unknown_fields))]
 pub struct BlockWithTxs {
     /// Status
@@ -107,7 +108,7 @@ pub struct BlockWithTxs {
     #[serde_as(as = "FeltHex")]
     pub new_root: Felt252,
     /// The time in which the block was created, encoded in Unix time
-    pub timestamp: u64,
+    pub timestamp: u128,
     /// The Starknet identity of the sequencer submitting this block
     #[serde_as(as = "FeltHex")]
     pub sequencer_address: Felt252,
